+++ love/src/Config.js
@@ -2945,6 +2945,14 @@ export const mtDomeTrackingStatetoStyle = {
   UNKNOWN: 'invalid',
 };
 
+export const mtDomeControlModeStateToStyle = {
+  UNKNOWN: 'invalid',
+  REMOTE: 'ok',
+  'LOCAL PUSH BUTTONS': 'alert',
+  'LOCAL KEBA': 'alert',
+  'LOCAL EUI': 'alert',
+};
+
 export const atDomeTrackingStatetoStyle = {
   UNKNOWN: 'invalid',
 };
@@ -3090,6 +3098,31 @@ export const schedulerBlocksStateToStyle = {
   COMPLETED: 'ok',
   ERROR: 'alert',
   INTERRUPTED: 'alert',
+};
+
+// POWER MONITOR
+export const POWER_MONITOR_STATUS_MAPPING = {
+  ONLINE: 'ok',
+  OFFLINE: 'alert',
+  RUNNING: 'ok',
+  STOPPED: 'invalid',
+  HEALTHY: 'ok',
+  'NO SOURCE': 'alert',
+  'ON BATTERY': 'alert',
+  'NO POWER': 'alert',
+  UNKNOWN: 'unknown',
+};
+
+export const POWER_MONITOR_CRITICALITY_MAPPING = {
+  HIGH: 'alert',
+  MEDIUM: 'warning',
+  LOW: 'ok',
+};
+
+export const POWER_MONITOR_STABILITY_MAPPING = {
+  STABLE: 'ok',
+  UNSTABLE: 'warning',
+  'HIGHLY UNSTABLE': 'alert',
 };
 
 /*****************************************************************************/
@@ -3321,6 +3354,14 @@ export const mtDomeMotionStateMap = {
   53: 'DISABLING',
   54: 'ENABLED',
   55: 'ENABLING',
+};
+
+export const mtDomeControlModeStateMap = {
+  0: 'UNKNOWN',
+  1: 'REMOTE',
+  2: 'LOCAL PUSH BUTTONS',
+  3: 'LOCAL KEBA',
+  4: 'LOCAL EUI',
 };
 
 export const atDomeTrackingStateMap = {