+++ love/src/components/ScriptQueue/ScriptQueue.jsx
@@ -554,7 +554,7 @@ export default class ScriptQueue extends Component {
     return !allowedCommands.includes(commandName);
   };
 
-  observatoryStateCommand = (newState, note) => {
+  observatoryStateCommand = (newState, note, sendNarrativelog = false) => {
     this.props.requestSALCommand(
       {
         csc: 'Scheduler',
@@ -565,17 +565,18 @@ export default class ScriptQueue extends Component {
         },
       },
       (statusCode, ackData) => {
-        if (statusCode === 200) {
+        if (statusCode === 200 && sendNarrativelog) {
           const observatoryStatusLabels = getActiveObservatoryStates(newState)
             .map((status) => OBSERVATORY_STATE_DETAIL[status]?.name ?? status)
             .join(' | ');
           const messageText =
             `Observatory status updated to: ${observatoryStatusLabels}.\n` + `Note: ${note ?? 'Not provided.'}`;
           const now = new Date();
+          const nowTai = new Date(now.getTime() - this.props.taiToUtc * 1000); // this.props.taiToUtc is negative
           ManagerInterface.createMessageNarrativeLogs({
             level: 0,
-            date_begin: now.toISOString().slice(0, -1),
-            date_end: now.toISOString().slice(0, -1),
+            date_begin: nowTai.toISOString().slice(0, -1),
+            date_end: nowTai.toISOString().slice(0, -1),
             message_text: messageText,
             is_human: true,
             category: 'None',