+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Version History
 ===============
 
+v6.12.0
+-------
+
+* Update package.json version to v6.12.0. `<https://github.com/lsst-ts/LOVE-frontend/pull/798>`_
+* Include MTDome controlMode event in the MTDome component. `<https://github.com/lsst-ts/LOVE-frontend/pull/797>`_
+* Implement PowerMonitor component. `<https://github.com/lsst-ts/LOVE-frontend/pull/796>`_
+* Add a toggle to the ObservatoryStatusMenu component to control whether to create a narrative log entries on observatory status transitions triggered by users. `<https://github.com/lsst-ts/LOVE-frontend/pull/795>`_
+
 v6.11.4
 -------